Add files via upload
In this project color yellow is detected. First, the webcam has been setup and image is converted to HSV color from BGR. Then lowerlimit and upper limit is setup so that to create mask and eventually find  the contours. When contours are found then bounding rectangle is applied on it in green color.